Nhat fix exercise ss05

+++ ss05_orm_jpa/exercise/product_management_jpa/src/main/java/com/example/model/Product.java
@@ -11,14 +11,14 @@ public class Product {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
     private String name;
-    private int price;
+    private double price;
     private String description;
     private String manufacturer;
 
     public Product() {
     }
 
-    public Product(int id, String name, int price, String description, String manufacturer) {
+    public Product(int id, String name, double price, String description, String manufacturer) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -46,11 +46,11 @@ public class Product {
         this.name = name;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
